+++ shared/loan/src/main/java/com/vtorushin/shared/loan/domain/entity/LoanCondition.kt
@@ -1,7 +1,7 @@
 package com.vtorushin.shared.loan.domain.entity
 
 data class LoanCondition(
-    val maxAmount: Int,
+    val maxAmount: Long,
     val percent: Double,
     val period: Int
 )
